Unify message image loading behind ImageResource protocol
